Retire FileAttachmentUploadService and refactor FileAttachmentBlobService
This removes the FileAttachmentUploadService and instead handles the
concerns of that in a create_initial method on FileAttachment::Revision.
This is to reflect that FileAttachmentUploadService is not performing
any uploading and thus it's name had become somewhat irrelevant.

It refactors the FileAttachmentBlobService to act more like a service
that performs an action and returns a result, as opposed to a model
where various attributes are looked up.

The FileAttachmentBlobService class has been through a few iterations
where it returned a hash and has now changed to return a
FileAttachment::BlobRevision object - this is something that is now
possible with the recent change to the FileAttachmentRevisionUpdater.