Don't treat blob as a sub revision for file attachments
It is useful to allow replacing a blob_revision on a
FileAttachmentRevisionUpdater as this is an entity that changes as a
whole rather than by individual attributes. This allows the change of
this to happen as a whole rather than passing attributes in.

Previously we could set a blob_revision but it would then create a
situation where the updater was out of sync with the sub revision
updater. This change removes that possibility.