Change FileAttachment asset to be a one-to-one relationship
Previously we expected file attachments to potentially multiple assets
associated with each one, as is the case with images. This was based on
the expectation that PDFs would have a file and a thumbnail asset.
However later it was decided that PDF preview images would not be a part
of Content Publisher.

This therefore simplifies the model code to only deal with a single
asset association for a FileAttachment::BlobRevision. This allows for
numerous methods to be simplified to avoid a loop.

The owner of the association remains to be the Asset model, this allows
us to switch back to the old one-to-many approach without too much
effort. However this comes at the expense that we can't guarantee an
asset existence by a foreign key, therefore instead we use the Rails
required method to validate the existence.