Remove variant column from file_attachment_assets
As there are no longer variants for file_attachment_assets this column
is no longer needed.

The blob_revision_id index is also updated to be a unique index to stop
the same asset referring to multiple blob_revisions, this stops this
becoming a one-to-many relationship with a database constraint.