Add a publishing requirement to select a primary organisation
The Primary Organisation field can now accept a blank string, however using
a new Tag Checker we insist that the user must submit a valid organisation
not only to publish the document, but also to save the Edit Tags page.

Browser no longer in a redirect loop if you have a nil organisation and visit the home page.

Now pass a revision to Edit Tags view instead of an edition to maintain the user's previous form submission.

Moved the point where we fail until after the issues are checked to make sure validation errors were found even with no changes made.