Refactor summary component variables
This changes the input to the summary component previously there was a
title section when included edit_url fields. This now a single field for
title and a separate block for edit which can take href and
data_attributes values.

Inside each item there is now an edit block that can take href and
data_attributes values.

This was done so data_attributes could be kept in close proximity to
other link related data since they are displayed on the same element.

The href attribute of edit objects uses the fetch method to ensure an
error if a href is not given.