Remove scheduled publishing datetime hidden tag
By checking if a scheduled publishing datetime is present on an edition
already, we can determine whether it is being updated or set for the first
time. This means we no longer needed a hidden tag in the scheduled publishing
datetime form to determine if it's an update or not.