Creates notification when a scheduled edition is published
This commit adds the ability to send a notification when a scheduled
document is published. A lot of the notification code here is similar to
when a unscheduled document is published but I chose to create a
seperate mailer and template etc to avoid filling the code with lots of
conditionals.

I have also added a spec for the `ScheduledPublishMailer` as this is the
only place we could test the output of the actual email. `PublishMailer`
achieves this by testing in the publish_spec feature which isn't really
an option for the publishing of a scheduled edition workflow.