Use updated ScheduledDateTimeChecker
This updates the controller and view code to make use of the updated
interface to the ScheduledDateTimeChecker.

This sets up links to errors that have occurred and shows the error for
the scheduled_date field. Unfortunately this comes with a couple of
issues:

- the date_input component does not support error items, so we have the
  array to string conversion;
- the label ends up outside of the error block.

This does not plug in an error for the time field as the time field
needs to be updated to cater for this.

This all makes some Time.zone.now into Time.current wherever I saw them.