Report page name for all events
Previously we defined data-gtm tags in order to support robust analytics
for user actions, as more volatile things like CSS and link text change
over time. When a user action can occur on multiple pages, such when in
the context of a modal, it can be difficult to filter by workflow. This
adds an automated 'page-name' attribute to make it easier to filter by
workflow when the same user actions can occur on multiple pages.