Add initial value data attribute to autocomplete
This allows us to specify a value that is outside of the underlying
select element for the autocomplete.

This is only of use with the autocomplete_without_narrowing_results type
as this enhances a select into an input and allows a user to enter free
text there. This causes problems if the user is re-presented the form as
we need to then put this data into the select list, which is especially
weird if the user has typed junk in.