Change approach to user entered times
Previously we allowed any user input to be part of the autocomplete
selections for a time input. This could produce some rather strange
results when a user didn't enter the time in a format we recognise.

To counter this problem this commit changes the select to only have
an extra value from a date that content publisher provides rather than
the user. User input is set with the initial value attribute on the
autocomplete so it is still available.

This refactors the code for the time options as the previous one didn't
appear to be working, so I tried to rewrite it in a simple, more concise
way.