Be more forgiving with date input
This allows users to enter a wider range of valid dates than just our
exact format of %-l:%M%P

Users are required to enter an hour and a minute, with an optional
period - which is separated by a optional space. Without a period we
take the hour literally so 9:34 is treated as 9:34am and 12:00 is
treated as 12:00pm.