Create Clear Datetime interactor
This extracts the logic for clearing a scheduled publishing datetime into an
interactor.  We have made a content decision that we no longer need to create a
timeline entry when a scheduled publishing datetime is clearer as this is an
edit. This is consistent with not creating a timeline entry when a scheduled
publishing datetime is updated.  We have also decided not to check if a
scheduled publishing datetime is present before clearing it, as if it's not
present the updater won't assign a new revision as it won't have changed.