Refactor schedule routes
Previously we had a variety of route-styles and naming for schedule
actions. This refactors the routes into two resource groups:
'schedule-proposal' and 'schedule', making the corresponding changes to
the controllers, views, actions, interactors, locales and specs.

Since the route to update an existing schedule is now separate, and the
original was not fully implemented (it raised a 500), this also removes
the associated link from the view, which has the advantage that we can
start from scratch on this feature.