Fix back link from the schedule page
Previously the 'Back' link from the page to schedule a document would go
to the document summary, which wasn't appropriate when coming from the
page to set a proposed date/time. Because this behaviour depends on the
previous page and not on the state of the document, we need a 'wizard'!

This modifies parts of the schedule workflow to add a 'wizard' query
param, which is then used to modify the 'Back' behaviour on the page to
schedule a document, and to replace the previous logic about whether to
check an action had been selected on the schedule proposal edit page.