Support rescheduling in the ScheduleService
Previously we created the ScheduledPublishingJob, which contains a few
checks that (should) make it a no-op when an edition is rescheduled. The
act of rescheduling an edition is therefore as simple as enqueueing
another ScheduledPublishingJob with the updated publishing time.

When an edition is rescheduled, the previous scheduling state should be
maintained, but the action could be performed by a different user and
also represents a distinct event in the document history; much of this
code is already present in the ScheduleService, ready to reuse.

Previously the enqueueing of a ScheduledPublishingJob was handled by the
schedule CreateInteractor, which made it hard to connect the changes
made by the ScheduleService and the checks performed by the job. This
moves the enqueueing of the job inside the schedule service, so that we
can explicitly assert that another job is expected to be enqueued.