Replace Edition#schedulable? with inline checks
Previously we had a schedulable? method on each edition that was used
for sanity checking in a couple of controller actions, as well as to
modify the actions shown to the user. The latter behaviour conflicts
with the new design for schedule requirements, as the 'Schedule' action
would disappear if the proposed time drifted into the past.

This replaces the schedulable? check for user actions by simply checking
the scheduled_publishing_datetime, noting the edition is editable in
both scenarios. Since it is valid to enqueue a job in the past (?!),
this also inlines and simplifies the schedulable? check when scheduling
a document, and removes it when harmlessly rendering the form.