Revisit analytics for radio buttons
Previously we had a custom event listener for submitting a form with
radio buttons. While this gave us accurate submission data, it seems
redundant to have a bespoke data capture method for certain forms when
this and other data are available in the database for the app.

This replaces the previous listener with a data-gtm approach on the
individual radio buttons, the aim being to capture an event for every
radio selection made. One use of this data could be to monitor the
effectiveness of the options for e.g. creating a new document.