Add a publish_time field to Scheduling
This column is introduced so that a scheduling can store the time that
content is intended to be published. As this is a table that already
exists - but not used in production - two queries are used to create the
column, populate it with a default and then remove the default.

Previously this column wasn't needed as it's role was accomplished by using
the scheduled_publishing_datetime value on a MetadatRevision. However
the intention is now to store a proposed publishing time on a
MetadatRevision and a concrete publish time on the Scheduling model.

The reason this is done is to help handle failed scheduled publishings.
The UI intends to refer to the time these were supposed to be published
and if this time is stored on an edition we hit a problem if it is
changed (by the document being re-scheduled).

This change also has the nice benefit of de-coupling the time from the
revision once it is made concrete which should be good if we ever allow
reverting to edits as they won't carry the time with them.