Refactor date/time validation into parser
Previously we validated and semantically checked the proposed schedule
date/time in a single requirements checker, which also made the parsed
date/time available. This splits the parsing/validation aspect into a
new DatetimeParser class that enables the remaining checker to operate
on the revision, instead of the request params, which means we can
re-use it to show requirements issues when the date/time drifts into the
past.