Fix parsing of time
This was having problems with times around the hour of 12 in both am and
pm. This was due to the previous code (which I wrote 😭) being a bit too
simplistic with time handling and handling every hour but 12. There are
now extra tests to catch this.

I've also added some extra tests to make clear some of the time
scenarios that should be rejected, these are mostly types of parsing
that are more complex than this time should accept but `Time.parse`
would.