Update ScheduleService for publish time changes
This changes the ScheduleService to create a Scheduling and update an
Edition at the point of scheduling an edition to publish.

Previously an interactor would be updating the edition and then the
updated edition would go through the scheduling service, meaning the
edition was updated twice in a request. This isn't ideal as it leaves a
trail in the db. Doing this in one save treats this more as an atomic
operation.

To achieve changing revision and status at the same time the
assign_status method was needed to be updated on an edition so it could
use an unsaved revision.