Use 'gtm-action' attribute to specify action
Previously we used the 'Click Text' variable in GTM as the action for
all 'Data GTM' events. This can be voluminous (for forms, where this
variable is all the text in the form) and doesn't cover input elements,
where the value attribute/property is also unhelpful e.g. for UUIDs.

This introduces a 'data-gtm-action' attribute to manually specify the
action field for all 'Data GTM' events. Events without this attribute
will not have any action text, which should be covered by the category
field, which is specified by the value of the 'data-gtm' attribute.