Convert application to use publish_time fields
This does various changes throughout the application to make use of
proposed_publish_time and publish_time fields instead of
scheduled_publishing_datetime.

There are a variety of small functionality changes to handle this, such
as no longer creating revisions in the schedule update interactor, and
creating a new revision as part of unscheduling (to re-set a
proposed_publish_time value).

There is some clean-up that should be done in the tests for terminology
around this, but I'll do that in a separate commit.