Refactor scheduling tests for publish time terminology
This makes various amends to all of the scheduling feature tests due to
the change in terminology for publish time and proposed publish time.

A common change is the use of TimeHelpers to freeze time. This was done
in places where we had a mix of a programmatically generated dates and
hard coded times. For logic simplicity I felt it was simple to just
freeze the time and hard code them both.

This also reduces the number of tests to avoid duplication and does
various other renaming of things for consistency.