Add a failed_to_publish status
This status has been created for handling a scheduled publishing that
fails. It is expected that once retry attempts are exhausted an edition
will be transitioned to this state.

The details association of this status is expected to be a scheduling
model. This is the only part of this status that is coupled to
scheduling.