Rename warning summary component to inset prompt
This renaming is to reflect that the purpose of this component is going
to become broader than warnings as an error modifier is to be
introduced.

Inset prompt was chosen as a name as this component is very similar to
the inset text component [1]. As this has more of an emphasis on prompting
a user to make an action than the inset-text component I felt that was an
appropriate second word.

[1]: https://govuk-publishing-components.herokuapp.com/component-guide/inset_text