Add error option to inset-prompt component
This is to achieve the alert required for a failed publishing.

I wasn't too sure on the best way to get the background colour to meet
the design as there wasn't a variable set. I've used the
govuk-tint helper to get a colour that looks very close to the design.