Send scheduled publishing failed emails with service
This moves the notifications of a failed scheduling publishing to be a
concern of the ScheduledPublishingFailedService.

This is our first service to send notifications so it is slightly
unusual in this regard. However, there aren't clear approaches defined
for services (they vary quite substantially) and anything removed from
here just means more code in the job.