Add integration tests for scheduled publishing
This creates tests for the success and failure scenarios of
scheduled publishing. These are created as integration tests rather than
feature tests as this is a hard thing to test as a feature (jobs run
later in time).

It removes the spec for ScheduledPublishingFailedService since this is
covered by the integration test.