Remove gtm-doc-status and gtm-doc-type attributes
Previously we had two custom GTM attributes to communicate the visible
document status and type, although the latter was never actually used.
Both can now be inferred from other events, but we should ideally use
the app DB as the source of this kind of non-eventful information.

This removes both attributes and the associated field of the metadata
component, as it is now redundant.