Make ScheduleChecker more generic
This changes the ScheduleChecker to operate on a DateTime instance
rather than on an revision. This is to allow doing these checks in both
a proposing and updating contexts for scheduling.

To reflect this change I've renamed this to be PublishTimeChecker since
the input is variations on a publish_time.