Raise on state/reqs before schedule/publish
This adds a couple of 'raise' statements before we actually publish or
schedule a document. Although it's highly unlikely a user would be able
to get to the point of submitting the associated forms, it makes sense
to apply the state/requirements checks at these critical points.

Previously we added support for users accidentally publishing the same
edition twice (e.g. when using the app in multiple tabs). This approach
is inconsistent with all other features, so we should avoid it until we
can adopt a more thorough and complete strategy to manage these cases.