Remove capybara-chromedriver-logger gem
I accidentally removed this from the test suite by commenting out the
line. That was done because this gem does not work with newer versions
of Chrome that default to using the W3C bridge for communication.

Since we've been relying on a fork of this gem for a long time and it
now causes problems it's a sign that we can probably live without it.