Track flashes separately from requirements issues
Previously we used a single data-gtm attribute to capture the entire
text of each (error) flash, which is excessive and makes it difficult
to pick out individual requirements issues. This enables tracking for
individual requirements issues and makes them separate from the events
for the visibility of different flashes, where the title of the flash
should be sufficient for analytics.