Create DateParser
In order to be consistent with the DatetimeParser, we extract out date parsing
from the BackdateChecker into a DateParser class. We had initially hoped that
the DatetimeParser would be able to take care of date parsing for us, however
it requires a time in order to be able to do so. Users only provide a date when
backdating content and we would reasonably expect to receive a parsed Date
object back from the service rather than a Time object, which is what
DatetimeParser currently provides.

In the course of this work we've identified that the DatetimeParser is
currently coupled to scheduling when creating issues. Therefore we have
followed the interface built for DateParser and modified DatetimeParser to
accept a `issue_prefix` argument, therefore making it more generic and reusable
for future datetime parsing needs.