Make DateParser and DatetimeParser idempotent
Both DateParser and DatetimeParser require `#parse` to be called before
`#issues` in order to check for issues and populate `issue_items`, meaning that
it's not idempotent. This moves `issue_items` into `#parse` and ensures that a
default empty array is passed to CheckerIssues in the event that `#parse` is
not called before `#issues` is.