This adds a spec for the PublishMailer
As the ScheduledPublishMailer needs a mailer spec, we thought it would
be a good idea to include a PublishMailer spec. This also means we can
make the feature for publishing simpler by removing alternative publishing
scenarios which actually only tested the different variations of the
email (now covered by the tests in the publish_mailer_spec).

Trello:
https://trello.com/c/uC28c63Y/949-tie-up-scheduling-loose-ends