Refactor autocomplete component to allow input elements
This allows the autocomplete component to work with both input elements
and select elements. This has been done to resolve the somewhat strange
scenario where for a time input where we allow free text but fall back
to a select element. To achieve this required a variety of
customisations and some oddities (like setting the user input value via a
data-attribute).

This changes the initialisation of an autocomplete component to accept
either a `select` or `input` hash which has dedicated arguments for
each.

`select` supports arguments of options, selected, size, and multiple.

`input` supports arguments of value and options. The options are
rendered as a datalist for the input element.

This also updates the examples to show more variety and have more
descriptions.