Use a text input for time autocomplete
This is a more appropriate element for this input, it allows us to set
the input to whatever the user entered and not have to worry about
merging the users input inside the list of options.

It also provides a nicer experience for a user with JavaScript as the
initialisation from an input is less jarring than it is for a select
element.