More consistent app-side__action padding
This removes a persistent margin-bottom from app-side__actions box,
which doesn't seem to have a particular use and makes the padding on
this box bottom heavy.

To consistently make the padding even we need to remove the margins from
buttons when they are the last item in the actions. This can be
complicated because the buttons could be embedded in forms, which makes
the usage of :last-child tricky.

There is also a margin removed on .govuk-link elements which isn't
needed since it only has an effect when this is floated.