Only send filter events that have values
Previously we dispatched visibility events for all filters on the index
page, but then applied a selector in GTM to only really trigger if the
'data-gtm-action' attribute was set. As we start to use this attribute
for other elements, it's become normal to fire visibility events without
an associated 'action' e.g. for guidance. This pushes the special case
behaviour for these four elements into the codebase, using a question
method to determine if each has a 'value' and thus if we want to track.