Rename Scheduling Helper to Date Time Helper
We rename Scheduling Helper and its methods to be more generic so it can be
reused across the various places in the app where we format how dates and times
are displayed. We've also taken the decision to enforce the '-' flag for both
date and time to ensure that they are not padded. This was done inconsistently
throughout the app and there doesn't seem to be a clear reason why some dates
and times should be padded and some not.