Fix tracking attributes on email links
Previously we only used the utm_content query param on the in-app links
we send out in emails, which is not sufficient for tracking the use of
these links in GA. This replaces utm_content with three params:

   - utm_source: this is what we put in utm_content previously
   - utm_medium: always 'email'
   - utm_campaign: always 'govuk-publishing'

The last param is optional, but potentially useful to distinguish other
links to our app that populate the other two parameters, which would
pollute our analytics. Since the text of the link is now quite long,
this replaces the original text with a markdown link approach to
completely hide the query params, although they will still be shown in
the text variant of the email.