Log the error a scheduled publishing fails on
The retry_on block will prevent exceptions working their way up the
stack. This means that errors are not reported to sentry by default.

This calls `GovukError.notify` to notify sentry for the reason the scheduling
failed. This will allow us to determine whether we need to fix code or handle
the exception. If the failure is due to an intermittent error (such as
timeout) this wouldn't make it to Sentry as they are filtered out in
govuk_test config.