Clear the set backdate if content has not been published
To allow the user to remove a backdate that has been applied to some content,
this commit does the following:

- Creates a Destroy Interactor class and associated route for the Backdate, tied into
the controller, to allow clean removal of the backdate value from the Edition.
The backdate is now set to Nil when prompted.

- Adds a user action on the Edit Backdate page to allow the backdate to be removed
if there is one set. This action is not present if no date has been set.

- Adds an entry in the history tab for when a backdate is removed. Wording to be
confirmed. Along these lines, fixed a bug wherein we were referring to the backdate
from the Edition directly rather than an appropriate Revision. Fix will make more
calls to the database but prevents a Nil error when we nullify the backdate value.

- Adds an error for if the user hits an issue when editing the backdate, similar to
scheduling.