Remove broken/unwanted image upload link
Previously we had code to render a link to upload an image directly from
the summary, which would act the same as clicking 'Change'. Now that the
'Change' link is renamed to 'Edit', we don't believe the additional link
is necessary. It's also been broken for ages. This removes it!