Initial modelling for access limits
This creates an access limit model that is associated with an edition.

The access limit model stores:

- limit_type: whether all organisations or just primary organisation
  apply
- edition: the edition the limit applies to
- revision_at_creation: which revision the edition represented at the
  point of access limiting to know which changes were limited and not
- created_by: the user who created it
- created_at: when it was created

---

We had a discussion within the team about what the appropriate approach
was to model access limiting. We considered whether it should be
something that was part of a Revision or be a model in it's own right.

We decided that it didn't make sense to be on a revision as an access
limit is not part of a documents content and, as the access limit has to
be removed at the point of publish, this would mean changing an edition
at the point of publishing which potentially happens without user
interaction. This violates a principal of a revision where they should
only be created as a result of a user action.

---

A tricky decision was deciding the best way to associate this with an
edition.

I considered two approaches:

1. Store an edition_id on access_limits and an access_limit_id on
   editions. This is similar to approach used for Status
2. Store an active flag on access_limits. Have edition look up the
   access_limits row matching edition_id and active. This is similar to
   how current and live editions are stored on a document.

After experimenting with both I opted to go for the first option. I felt
this was more idiomatic to Rails and, much to my surprise, it worked
with FactoryBot without any tweaking (I thought I'd hit a race condition
where it would try create an AccessLimit before the Edition and fail due
edition_id) being nil.

The thing I preferred about the 2nd approach that is that it avoided
having bi-directional associations which feel unusual in a database.
However the things that are gained that make me prefer the first
approach are:

- easier to update an access limit in memory, the flag approach requires
  immediate persistence;
- Feels more consistent with other associations on an Edition such as
  Status and Revision.