Add methods to assign/remove access limits
I spent a while debating how to approach these methods as I wanted to
maintain the consistency of using an assign_ prefix to convey which
methods assigned attributed as opposed to autosaving. I couldn't think
of a good way to approach this for the method to remove access limit.

Eventually I concluded that since ActiveRecord doesn't have an approach
for a remove assign and since we don't already use the remove prefix
that this could be used without causing too much confusion.