Show Content Data URL when document is live
Previously we were only able to show a link to a document's Content Data URL
under certain conditions, otherwise users would be directed to a 404 page. This
issue has since been resolved in the Content Data app so we remove the logic
for displaying the link as it's no longer necessary. Now, we show the link when
a document has a live edition.