Add or replace assertions for all user actions
Previously we had some ad-hoc 'raise' assertions in our interactors.
This replaces the existing assertions and adds missing ones for all user
actions by including EditionAssertions for controllers and interactors.

Most of the assertions are self-explanatory. Ideally we would have most
assertions in controllers, since we would also like to handle them at
this point. For actions that modify a document, the assertions can only
go in the interactors because the edition is not available beforehand.

Some unusual cases:

   * Editing topics is technically possible on a live edition, but the
assertion behaviour matches that for the UI
   * Editing a withdrawal re-uses the same controller actions to
withdraw, so we need to support this use case
   * Downloading a lead image makes no change to the edition, so the
assertion is mostly to document when we expect this feature to be used