Redirect when a edition assertions fail
Previously we let failed assertions propagate into 500 errors. This
changes the approach to redirect to the document summary when edition
assertions fail. Examples of when this might happen are when a user is
working on an edition in multiple tabs (and forgets to refresh), or when
multiple users are working on the same edition over a period.

One special case is when the first edition is discarded, in which case
there is no edition to redirect to, so we redirect to the index page.

Another special case is the modal. Here we take a simpler approach that
causes the modal to display an error, with the intention that the user
will eventually refresh the page and trigger the normal assertion
redirect. Note that this relies on the current situation where no modal
action involves an action-specific behaviour when the assertion fails.