Move unwithdraw service into interactor
Previously we had a simple, single-use UnwithdrawServic - a legacy of
the initial rake-task-based support for this feature. This merges the
service into the associated interactor and tidies-up the feature specs.

One aspect of unwithdrawing an edition is to preserve its previously
published state. The feature test now checks this implicitly by setting
a 'non obvious' initial state of 'published but needs 2i'.