Add tests for error pages
This adds tests for error pages which check the correct response and
some of the correct content is returned. I felt that basic tests were
useful for simple checks for syntax errors and that translation keys
exist and are used.

Rather controversially these are request specs and the first ones in
Content Publisher. I felt request spec was the right choice as these are
not features, and wouldn't feel right as a type of those. They could
have been controller tests (which we also have none of) but these are
now discouraged [1].

[1]: http://rspec.info/blog/2016/07/rspec-3-5-has-been-released/#rails-support-for-rails-5