Allow removing of access limited editions
GOV.UK has a precedence to removing content that is access limited as
part of the process of syncing data from the production environment to
integration. This means that the access limited content is not available
in the integration environment or for GOV.UK developers cloning the
data.

This is achieved by the use of SQL scripts that are executed by the
[env-sync-and-backup](https://github.com/alphagov/env-sync-and-backup)
repository. It is not an ideal process by any means as it's decoupled
from applications, fragile to schema changes, difficult to test, and
near impossible to remove everything potentially sensitive.

To allow for this to operate we need to allow the ability to delete
records from the database, this was something prevented by default as
explained in [ADR-8][]. Now we have an actual need to perform this
operation this sets cascades such the deletion of an edition is possible
and any data that that depends on an edition is deleted by cascade.

This SQL script is used to delete any access limited editions and to
reset associated documents to refer to the previous item as the current
item, so that there is still a db record for the edition.

[ADR-8]: https://github.com/alphagov/content-publisher/blob/5966c344fffef043dba127fe5aa74a0bfdf7f184/docs/adr/0008-restricted-deletion.md