Allow revision deletion for access limiting
When access limited editions are removed as part of the production to
integration data-sync there are a number of orphaned records left that
contain the sensitive data we wish to remove.

To remove these there are a number of queries set-up that remove
revisions and related artefacts if there is not parent records
associated with them.

As part of this the on_delete operations for records that depend on
revisions are updated so a delete can cascade these. These are set-up
with the expectation that any associated editions for the revision is
already removed thus you could only delete a revision after an
associated edition is already removed.