Improve handling of nil tags/organisations
Previously we used '[]' to retrieve the 'internal_name' of tags and
organisations, which raises an error if the retrieved hash is nil. While
we don't expect this to happen in production, it can occur in
development without a complete data import; in general, we also don't
want to raise errors about what is essentially a content issue. This
replaces '[]' with '#fetch' to gracefully cope with nils.