Move uploading assets into preview service
This moves the method to upload an asset, which was previously in the
AssetManagerService, into the PreviewAssetService, which is its single
point of use. Having the upload and update calls in the same class means
we can more easily share a common payload between them to support access
limit settings.