Move publish/redirect calls to PublishAssetService
Previously we delegated publish/redirect to the AssetManagerService,
even though they are only used once. This moves the implementation of
these methods to the class where they are used.

Although we have not lost coverage, it looks like our testing of the
publish/redirect behaviour is incomplete, as the test helpers for Asset
Manager make it hard to check the body of the request.

We should aim to address this in future work.