Automatically enforce access limiting feature
Previously we had to remember to add access limit assertions for each
new user action where they apply - we've already missed a few. This
replaces the individual assertions with a global before_action:

   - We use 'request.path_parameters' to work out if the URL is trying
to access an edition under the 'documents/:document' scope
   - We replace the assertion/exception mechanism with a boolean check,
so that all the access limit code is in one place

While changing the mechanism may be pragmatic to overcome the issues
with repetition and forgetfulness, we recognise that it also has a few
downsides, which may cause pain in the future:

   - We now retrieve an edition twice in order to process a request
   - Our approach to enforcing access limits is now inconsistent with
other kinds of assertion
   - We now implicitly rely on the 'document' scope in the routes

This also moves the 'AccessLimit#organisation_ids' method onto the
Edition class to better reflect the variety of data it uses, and to
avoid doing even more extra queries when an edition is access limited -
our current relationships prevent using inverse_of memoisation.